Struct esp32c3::extmem::RegisterBlock

#[repr(C)]
pub struct RegisterBlock { /* private fields */ }

Register block

Implementations

impl RegisterBlock

pub const fn icache_ctrl(&self) -> &ICACHE_CTRL

0x00 - This description will be updated in the near future.

pub const fn icache_ctrl1(&self) -> &ICACHE_CTRL1

0x04 - This description will be updated in the near future.

pub const fn icache_tag_power_ctrl(&self) -> &ICACHE_TAG_POWER_CTRL

0x08 - This description will be updated in the near future.

pub const fn icache_prelock_ctrl(&self) -> &ICACHE_PRELOCK_CTRL

0x0c - This description will be updated in the near future.

pub const fn icache_prelock_sct0_addr(&self) -> &ICACHE_PRELOCK_SCT0_ADDR

0x10 - This description will be updated in the near future.

pub const fn icache_prelock_sct1_addr(&self) -> &ICACHE_PRELOCK_SCT1_ADDR

0x14 - This description will be updated in the near future.

pub const fn icache_prelock_sct_size(&self) -> &ICACHE_PRELOCK_SCT_SIZE

0x18 - This description will be updated in the near future.

pub const fn icache_lock_ctrl(&self) -> &ICACHE_LOCK_CTRL

0x1c - This description will be updated in the near future.

pub const fn icache_lock_addr(&self) -> &ICACHE_LOCK_ADDR

0x20 - This description will be updated in the near future.

pub const fn icache_lock_size(&self) -> &ICACHE_LOCK_SIZE

0x24 - This description will be updated in the near future.

pub const fn icache_sync_ctrl(&self) -> &ICACHE_SYNC_CTRL

0x28 - This description will be updated in the near future.

pub const fn icache_sync_addr(&self) -> &ICACHE_SYNC_ADDR

0x2c - This description will be updated in the near future.

pub const fn icache_sync_size(&self) -> &ICACHE_SYNC_SIZE

0x30 - This description will be updated in the near future.

pub const fn icache_preload_ctrl(&self) -> &ICACHE_PRELOAD_CTRL

0x34 - This description will be updated in the near future.

pub const fn icache_preload_addr(&self) -> &ICACHE_PRELOAD_ADDR

0x38 - This description will be updated in the near future.

pub const fn icache_preload_size(&self) -> &ICACHE_PRELOAD_SIZE

0x3c - This description will be updated in the near future.

pub const fn icache_autoload_ctrl(&self) -> &ICACHE_AUTOLOAD_CTRL

0x40 - This description will be updated in the near future.

pub const fn icache_autoload_sct0_addr(&self) -> &ICACHE_AUTOLOAD_SCT0_ADDR

0x44 - This description will be updated in the near future.

pub const fn icache_autoload_sct0_size(&self) -> &ICACHE_AUTOLOAD_SCT0_SIZE

0x48 - This description will be updated in the near future.

pub const fn icache_autoload_sct1_addr(&self) -> &ICACHE_AUTOLOAD_SCT1_ADDR

0x4c - This description will be updated in the near future.

pub const fn icache_autoload_sct1_size(&self) -> &ICACHE_AUTOLOAD_SCT1_SIZE

0x50 - This description will be updated in the near future.

pub const fn ibus_to_flash_start_vaddr(&self) -> &IBUS_TO_FLASH_START_VADDR

0x54 - This description will be updated in the near future.

pub const fn ibus_to_flash_end_vaddr(&self) -> &IBUS_TO_FLASH_END_VADDR

0x58 - This description will be updated in the near future.

pub const fn dbus_to_flash_start_vaddr(&self) -> &DBUS_TO_FLASH_START_VADDR

0x5c - This description will be updated in the near future.

pub const fn dbus_to_flash_end_vaddr(&self) -> &DBUS_TO_FLASH_END_VADDR

0x60 - This description will be updated in the near future.

pub const fn cache_acs_cnt_clr(&self) -> &CACHE_ACS_CNT_CLR

0x64 - This description will be updated in the near future.

pub const fn ibus_acs_miss_cnt(&self) -> &IBUS_ACS_MISS_CNT

0x68 - This description will be updated in the near future.

pub const fn ibus_acs_cnt(&self) -> &IBUS_ACS_CNT

0x6c - This description will be updated in the near future.

pub const fn dbus_acs_flash_miss_cnt(&self) -> &DBUS_ACS_FLASH_MISS_CNT

0x70 - This description will be updated in the near future.

pub const fn dbus_acs_cnt(&self) -> &DBUS_ACS_CNT

0x74 - This description will be updated in the near future.

pub const fn cache_ilg_int_ena(&self) -> &CACHE_ILG_INT_ENA

0x78 - This description will be updated in the near future.

pub const fn cache_ilg_int_clr(&self) -> &CACHE_ILG_INT_CLR

0x7c - This description will be updated in the near future.

pub const fn cache_ilg_int_st(&self) -> &CACHE_ILG_INT_ST

0x80 - This description will be updated in the near future.

pub const fn core0_acs_cache_int_ena(&self) -> &CORE0_ACS_CACHE_INT_ENA

0x84 - This description will be updated in the near future.

pub const fn core0_acs_cache_int_clr(&self) -> &CORE0_ACS_CACHE_INT_CLR

0x88 - This description will be updated in the near future.

pub const fn core0_acs_cache_int_st(&self) -> &CORE0_ACS_CACHE_INT_ST

0x8c - This description will be updated in the near future.

pub const fn core0_dbus_reject_st(&self) -> &CORE0_DBUS_REJECT_ST

0x90 - This description will be updated in the near future.

pub const fn core0_dbus_reject_vaddr(&self) -> &CORE0_DBUS_REJECT_VADDR

0x94 - This description will be updated in the near future.

pub const fn core0_ibus_reject_st(&self) -> &CORE0_IBUS_REJECT_ST

0x98 - This description will be updated in the near future.

pub const fn core0_ibus_reject_vaddr(&self) -> &CORE0_IBUS_REJECT_VADDR

0x9c - This description will be updated in the near future.

pub const fn cache_mmu_fault_content(&self) -> &CACHE_MMU_FAULT_CONTENT

0xa0 - This description will be updated in the near future.

pub const fn cache_mmu_fault_vaddr(&self) -> &CACHE_MMU_FAULT_VADDR

0xa4 - This description will be updated in the near future.

pub const fn cache_wrap_around_ctrl(&self) -> &CACHE_WRAP_AROUND_CTRL

0xa8 - This description will be updated in the near future.

pub const fn cache_mmu_power_ctrl(&self) -> &CACHE_MMU_POWER_CTRL

0xac - This description will be updated in the near future.

pub const fn cache_state(&self) -> &CACHE_STATE

0xb0 - This description will be updated in the near future.

pub const fn cache_encrypt_decrypt_record_disable( &self ) -> &CACHE_ENCRYPT_DECRYPT_RECORD_DISABLE

0xb4 - This description will be updated in the near future.

pub const fn cache_encrypt_decrypt_clk_force_on( &self ) -> &CACHE_ENCRYPT_DECRYPT_CLK_FORCE_ON

0xb8 - This description will be updated in the near future.

pub const fn cache_preload_int_ctrl(&self) -> &CACHE_PRELOAD_INT_CTRL

0xbc - This description will be updated in the near future.

pub const fn cache_sync_int_ctrl(&self) -> &CACHE_SYNC_INT_CTRL

0xc0 - This description will be updated in the near future.

pub const fn cache_mmu_owner(&self) -> &CACHE_MMU_OWNER

0xc4 - This description will be updated in the near future.

pub const fn cache_conf_misc(&self) -> &CACHE_CONF_MISC

0xc8 - This description will be updated in the near future.

pub const fn icache_freeze(&self) -> &ICACHE_FREEZE

0xcc - This description will be updated in the near future.

pub const fn icache_atomic_operate_ena(&self) -> &ICACHE_ATOMIC_OPERATE_ENA

0xd0 - This description will be updated in the near future.

pub const fn cache_request(&self) -> &CACHE_REQUEST

0xd4 - This description will be updated in the near future.

pub const fn ibus_pms_tbl_lock(&self) -> &IBUS_PMS_TBL_LOCK

0xd8 - This description will be updated in the near future.

pub const fn ibus_pms_tbl_boundary0(&self) -> &IBUS_PMS_TBL_BOUNDARY0

0xdc - This description will be updated in the near future.

pub const fn ibus_pms_tbl_boundary1(&self) -> &IBUS_PMS_TBL_BOUNDARY1

0xe0 - This description will be updated in the near future.

pub const fn ibus_pms_tbl_boundary2(&self) -> &IBUS_PMS_TBL_BOUNDARY2

0xe4 - This description will be updated in the near future.

pub const fn ibus_pms_tbl_attr(&self) -> &IBUS_PMS_TBL_ATTR

0xe8 - This description will be updated in the near future.

pub const fn dbus_pms_tbl_lock(&self) -> &DBUS_PMS_TBL_LOCK

0xec - This description will be updated in the near future.

pub const fn dbus_pms_tbl_boundary0(&self) -> &DBUS_PMS_TBL_BOUNDARY0

0xf0 - This description will be updated in the near future.

pub const fn dbus_pms_tbl_boundary1(&self) -> &DBUS_PMS_TBL_BOUNDARY1

0xf4 - This description will be updated in the near future.

pub const fn dbus_pms_tbl_boundary2(&self) -> &DBUS_PMS_TBL_BOUNDARY2

0xf8 - This description will be updated in the near future.

pub const fn dbus_pms_tbl_attr(&self) -> &DBUS_PMS_TBL_ATTR

0xfc - This description will be updated in the near future.

pub const fn clock_gate(&self) -> &CLOCK_GATE

0x100 - This description will be updated in the near future.

pub const fn reg_date(&self) -> &REG_DATE

0x3fc - This description will be updated in the near future.