Module esp32c3::rtc_cntl::ext_xtl_conf

rtc configure register

Structs

EXT_XTL_CONF_SPEC

rtc configure register

Type Aliases

DAC_XTAL_32K_R

Field DAC_XTAL_32K reader - DAC_XTAL_32K

DAC_XTAL_32K_W

Field DAC_XTAL_32K writer - DAC_XTAL_32K

DBUF_XTAL_32K_R

Field DBUF_XTAL_32K reader - 0: single-end buffer 1: differential buffer

DBUF_XTAL_32K_W

Field DBUF_XTAL_32K writer - 0: single-end buffer 1: differential buffer

DGM_XTAL_32K_R

Field DGM_XTAL_32K reader - xtal_32k gm control

DGM_XTAL_32K_W

Field DGM_XTAL_32K writer - xtal_32k gm control

DRES_XTAL_32K_R

Field DRES_XTAL_32K reader - DRES_XTAL_32K

DRES_XTAL_32K_W

Field DRES_XTAL_32K writer - DRES_XTAL_32K

ENCKINIT_XTAL_32K_R

Field ENCKINIT_XTAL_32K reader - apply an internal clock to help xtal 32k to start

ENCKINIT_XTAL_32K_W

Field ENCKINIT_XTAL_32K writer - apply an internal clock to help xtal 32k to start

R

Register EXT_XTL_CONF reader

W

Register EXT_XTL_CONF writer

WDT_STATE_R

Field WDT_STATE reader - state of 32k_wdt

XPD_XTAL_32K_R

Field XPD_XTAL_32K reader - XPD_XTAL_32K

XPD_XTAL_32K_W

Field XPD_XTAL_32K writer - XPD_XTAL_32K

XTAL32K_AUTO_BACKUP_R

Field XTAL32K_AUTO_BACKUP reader - xtal 32k switch to back up clock when xtal is dead

XTAL32K_AUTO_BACKUP_W

Field XTAL32K_AUTO_BACKUP writer - xtal 32k switch to back up clock when xtal is dead

XTAL32K_AUTO_RESTART_R

Field XTAL32K_AUTO_RESTART reader - xtal 32k restart xtal when xtal is dead

XTAL32K_AUTO_RESTART_W

Field XTAL32K_AUTO_RESTART writer - xtal 32k restart xtal when xtal is dead

XTAL32K_AUTO_RETURN_R

Field XTAL32K_AUTO_RETURN reader - xtal 32k switch back xtal when xtal is restarted

XTAL32K_AUTO_RETURN_W

Field XTAL32K_AUTO_RETURN writer - xtal 32k switch back xtal when xtal is restarted

XTAL32K_EXT_CLK_FO_R

Field XTAL32K_EXT_CLK_FO reader - xtal 32k external xtal clock force on

XTAL32K_EXT_CLK_FO_W

Field XTAL32K_EXT_CLK_FO writer - xtal 32k external xtal clock force on

XTAL32K_GPIO_SEL_R

Field XTAL32K_GPIO_SEL reader - XTAL_32K sel. 0: external XTAL_32K

XTAL32K_GPIO_SEL_W

Field XTAL32K_GPIO_SEL writer - XTAL_32K sel. 0: external XTAL_32K

XTAL32K_WDT_CLK_FO_R

Field XTAL32K_WDT_CLK_FO reader - xtal 32k watch dog clock force on

XTAL32K_WDT_CLK_FO_W

Field XTAL32K_WDT_CLK_FO writer - xtal 32k watch dog clock force on

XTAL32K_WDT_EN_R

Field XTAL32K_WDT_EN reader - xtal 32k watch dog enable

XTAL32K_WDT_EN_W

Field XTAL32K_WDT_EN writer - xtal 32k watch dog enable

XTAL32K_WDT_RESET_R

Field XTAL32K_WDT_RESET reader - xtal 32k watch dog sw reset

XTAL32K_WDT_RESET_W

Field XTAL32K_WDT_RESET writer - xtal 32k watch dog sw reset

XTAL32K_XPD_FORCE_R

Field XTAL32K_XPD_FORCE reader - Xtal 32k xpd control by sw or fsm

XTAL32K_XPD_FORCE_W

Field XTAL32K_XPD_FORCE writer - Xtal 32k xpd control by sw or fsm

XTL_EXT_CTR_EN_R

Field XTL_EXT_CTR_EN reader - enable gpio configure xtal power on

XTL_EXT_CTR_EN_W

Field XTL_EXT_CTR_EN writer - enable gpio configure xtal power on

XTL_EXT_CTR_LV_R

Field XTL_EXT_CTR_LV reader - 0: power down XTAL at high level

XTL_EXT_CTR_LV_W

Field XTL_EXT_CTR_LV writer - 0: power down XTAL at high level