Module esp32c3::spi1::flash_sus_cmd

SPI1 flash suspend command register

Structs

FLASH_SUS_CMD_SPEC

SPI1 flash suspend command register

Type Aliases

FLASH_PER_COMMAND_R

Field FLASH_PER_COMMAND reader - Program/Erase resume command.

FLASH_PER_COMMAND_W

Field FLASH_PER_COMMAND writer - Program/Erase resume command.

FLASH_PES_COMMAND_R

Field FLASH_PES_COMMAND reader - Program/Erase suspend command.

FLASH_PES_COMMAND_W

Field FLASH_PES_COMMAND writer - Program/Erase suspend command.

R

Register FLASH_SUS_CMD reader

W

Register FLASH_SUS_CMD writer

WAIT_PESR_COMMAND_R

Field WAIT_PESR_COMMAND reader - Flash SUS/SUS1/SUS2 status bit read command. The command should be sent when SUS/SUS1/SUS2 bit should be checked to insure the suspend or resume status of flash.

WAIT_PESR_COMMAND_W

Field WAIT_PESR_COMMAND writer - Flash SUS/SUS1/SUS2 status bit read command. The command should be sent when SUS/SUS1/SUS2 bit should be checked to insure the suspend or resume status of flash.