Type Alias esp32c3::spi1::user::R

pub type R = R<USER_SPEC>;

Register USER reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn ck_out_edge(&self) -> CK_OUT_EDGE_R

Bit 9 - the bit combined with spi_mem_mosi_delay_mode bits to set mosi signal delay mode.

pub fn fwrite_dual(&self) -> FWRITE_DUAL_R

Bit 12 - In the write operations read-data phase apply 2 signals

pub fn fwrite_quad(&self) -> FWRITE_QUAD_R

Bit 13 - In the write operations read-data phase apply 4 signals

pub fn fwrite_dio(&self) -> FWRITE_DIO_R

Bit 14 - In the write operations address phase and read-data phase apply 2 signals.

pub fn fwrite_qio(&self) -> FWRITE_QIO_R

Bit 15 - In the write operations address phase and read-data phase apply 4 signals.

pub fn usr_miso_highpart(&self) -> USR_MISO_HIGHPART_R

Bit 24 - read-data phase only access to high-part of the buffer spi_mem_w8~spi_mem_w15. 1: enable 0: disable.

pub fn usr_mosi_highpart(&self) -> USR_MOSI_HIGHPART_R

Bit 25 - write-data phase only access to high-part of the buffer spi_mem_w8~spi_mem_w15. 1: enable 0: disable.

pub fn usr_dummy_idle(&self) -> USR_DUMMY_IDLE_R

Bit 26 - SPI clock is disable in dummy phase when the bit is enable.

pub fn usr_mosi(&self) -> USR_MOSI_R

Bit 27 - This bit enable the write-data phase of an operation.

pub fn usr_miso(&self) -> USR_MISO_R

Bit 28 - This bit enable the read-data phase of an operation.

pub fn usr_dummy(&self) -> USR_DUMMY_R

Bit 29 - This bit enable the dummy phase of an operation.

pub fn usr_addr(&self) -> USR_ADDR_R

Bit 30 - This bit enable the address phase of an operation.

pub fn usr_command(&self) -> USR_COMMAND_R

Bit 31 - This bit enable the command phase of an operation.