Type Alias esp32c3::twai0::int_raw::R

pub type R = R<INT_RAW_SPEC>;

Register INT_RAW reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rx_int_st(&self) -> RX_INT_ST_R

Bit 0 - Receive interrupt. If this bit is set to 1, it indicates there are messages to be handled in the RX FIFO.

pub fn tx_int_st(&self) -> TX_INT_ST_R

Bit 1 - Transmit interrupt. If this bit is set to 1, it indicates the message transmitting mis- sion is finished and a new transmission is able to execute.

pub fn err_warn_int_st(&self) -> ERR_WARN_INT_ST_R

Bit 2 - Error warning interrupt. If this bit is set to 1, it indicates the error status signal and the bus-off status signal of Status register have changed (e.g., switched from 0 to 1 or from 1 to 0).

pub fn overrun_int_st(&self) -> OVERRUN_INT_ST_R

Bit 3 - Data overrun interrupt. If this bit is set to 1, it indicates a data overrun interrupt is generated in the RX FIFO.

pub fn err_passive_int_st(&self) -> ERR_PASSIVE_INT_ST_R

Bit 5 - Error passive interrupt. If this bit is set to 1, it indicates the TWAI Controller is switched between error active status and error passive status due to the change of error counters.

pub fn arb_lost_int_st(&self) -> ARB_LOST_INT_ST_R

Bit 6 - Arbitration lost interrupt. If this bit is set to 1, it indicates an arbitration lost interrupt is generated.

pub fn bus_err_int_st(&self) -> BUS_ERR_INT_ST_R

Bit 7 - Error interrupt. If this bit is set to 1, it indicates an error is detected on the bus.