Module esp32c3::uart0::rs485_conf

RS485 mode configuration

Structs

RS485_CONF_SPEC

RS485 mode configuration

Type Aliases

DL0_EN_R

Field DL0_EN reader - Set this bit to delay the stop bit by 1 bit.

DL0_EN_W

Field DL0_EN writer - Set this bit to delay the stop bit by 1 bit.

DL1_EN_R

Field DL1_EN reader - Set this bit to delay the stop bit by 1 bit.

DL1_EN_W

Field DL1_EN writer - Set this bit to delay the stop bit by 1 bit.

R

Register RS485_CONF reader

RS485RXBY_TX_EN_R

Field RS485RXBY_TX_EN reader - 1’h1: enable rs485 transmitter to send data when rs485 receiver line is busy.

RS485RXBY_TX_EN_W

Field RS485RXBY_TX_EN writer - 1’h1: enable rs485 transmitter to send data when rs485 receiver line is busy.

RS485TX_RX_EN_R

Field RS485TX_RX_EN reader - Set this bit to enable receiver could receive data when the transmitter is transmitting data in rs485 mode.

RS485TX_RX_EN_W

Field RS485TX_RX_EN writer - Set this bit to enable receiver could receive data when the transmitter is transmitting data in rs485 mode.

RS485_EN_R

Field RS485_EN reader - Set this bit to choose the rs485 mode.

RS485_EN_W

Field RS485_EN writer - Set this bit to choose the rs485 mode.

RS485_RX_DLY_NUM_R

Field RS485_RX_DLY_NUM reader - This register is used to delay the receiver’s internal data signal.

RS485_RX_DLY_NUM_W

Field RS485_RX_DLY_NUM writer - This register is used to delay the receiver’s internal data signal.

RS485_TX_DLY_NUM_R

Field RS485_TX_DLY_NUM reader - This register is used to delay the transmitter’s internal data signal.

RS485_TX_DLY_NUM_W

Field RS485_TX_DLY_NUM writer - This register is used to delay the transmitter’s internal data signal.

W

Register RS485_CONF writer